% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discernible.R
\name{discernible}
\alias{discernible}
\title{Replace statistically significant with statistically discernible}
\usage{
discernible(x)
}
\arguments{
\item{x}{character string.}
}
\description{
Replaces the phrase statistically significant with statistically discernible
in a given character string.
}
\examples{
conclusion <- "The clinical trial showed statistically significant evidence of being effective."
discernible(conclusion)
}
